requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
black==24.8.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
detectron2==0.6
dominate==2.9.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
fonttools==4.53.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
grpcio==1.65.4
hydra-core==1.3.2
idna==3.8
imageio==2.35.1
iopath==0.1.9
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.1
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
pycocotools==2.0.8
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
tifffile==2024.8.30
tomli==2.0.1
torch==2.4.0+cu124
torch-tb-profiler==0.4.3
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visitor==0.1.3
wcwidth==0.2.13
Werkzeug==3.0.3
yacs==0.1.8
